+++ Kiri2d/include/kiri2d/hdv_toolkit/primitives/vertex3.h
@@ -1,3 +1,10 @@
+/*** 
+ * @Author: Xu.WANG
+ * @Date: 2021-12-23 17:57:21
+ * @LastEditTime: 2021-12-29 14:58:25
+ * @LastEditors: Xu.WANG
+ * @Description: 
+ */
 /***
  * @Author: Xu.WANG
  * @Date: 2021-12-01 19:36:00
@@ -19,13 +26,13 @@ namespace HDV::Primitives
     public:
         explicit Vertex3() : Vertex(3) {}
         explicit Vertex3(int id) : Vertex(3, id) {}
-        explicit Vertex3(float x, float y, float z) : Vertex(3)
+        explicit Vertex3(double x, double y, double z) : Vertex(3)
         {
             mPosition[0] = x;
             mPosition[1] = y;
             mPosition[2] = z;
         }
-        explicit Vertex3(float x, float y, float z, int id) : Vertex(3, id)
+        explicit Vertex3(double x, double y, double z, int id) : Vertex(3, id)
         {
             mPosition[0] = x;
             mPosition[1] = y;
@@ -33,19 +40,19 @@ namespace HDV::Primitives
         }
         virtual ~Vertex3() noexcept {}
 
-        void Set(float x, float y, float z)
+        void Set(double x, double y, double z)
         {
             mPosition[0] = x;
             mPosition[1] = y;
             mPosition[2] = z;
         }
 
-        float Distance(float px, float py, float pz)
+        double Distance(double px, double py, double pz)
         {
             return std::sqrtf(SqrDistance(px, py, pz));
         }
 
-        float SqrDistance(float px, float py, float pz)
+        double SqrDistance(double px, double py, double pz)
         {
             auto x = mPosition[0] - px;
             auto y = mPosition[1] - py;
@@ -54,9 +61,9 @@ namespace HDV::Primitives
             return x * x + y * y + z * z;
         }
 
-        float X() const { return mPosition[0]; }
-        float Y() const { return mPosition[1]; }
-        float Z() const { return mPosition[2]; }
+        double X() const { return mPosition[0]; }
+        double Y() const { return mPosition[1]; }
+        double Z() const { return mPosition[2]; }
     };
     typedef std::shared_ptr<Vertex3> Vertex3Ptr;
 } // namespace HDV::Primitives
